:recycle: [refactory]: Update CustomerForm.tsx
- Add custom hook of useCustomerMutate.ts
- Update component of CustomerForm.tsx
